requirements.txt
watchdog>=3.0.0 
pywin32>=306 
tkcalendar 
pyminizip 
babel 
configparser 
pandas
openpyxl
xlrd
xlwt
PyPDF2
Pillow